+++ src/ballistica/shared/ballistica.cc
@@ -51,7 +51,7 @@ auto main(int argc, char** argv) -> int {
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kEngineBuildNumber = 22878;
+const int kEngineBuildNumber = 22880;
 const char* kEngineVersion = "1.8.0a1";
 const int kEngineApiVersion = 9;